[features test cmd] Play nice with setting 'remoteUser' in a test
Previously, if you set 'remoteUser' in a test, the accompanying test
scripts placed into the shared 'workspaceFolder' would not have the
appropriate permission bits set to be executed by the exec user.

This patch ensure that all files in the workspaceFolder can be
read/written/executed by the exec user.

A test is added to validate creating a user and setting 'remoteUser' in
a scenario, and then running a test within that context to validate.