Display the 'origin' of the lifecycle command in the output log during 'up'
Logs like 'Running the postCreateCommand from devcontainer.json'
are updated to include the origin of the command

e.g: 'Running the postCreateCommand from Feature 'common-utils'.

This patch adds in a mapping object that is passed along and used to
map a command to an origin.

This change is a bit less efficient, with the added benefit of not
needing to modify the existing merging logic.