Rework overrideFeatureInstallOrder property
Instead of trying to augment the graph with artifical edges,
any Features that are listed in the overrideFeatureInstallOrder
are assigned a priority. The priority is used at the end of a round
of dependency resolution to sort the Features in the order they
should be installed via the override property.

Only process non-legacy Features in first pass of overrideFeatureInstallOrder