Build as single executable application
This commit adds a build script to package the `devcontainer` CLI as a
single executable application using the new (and experimental) Node.js
feature of the same name.
https://nodejs.org/api/single-executable-applications.html

Distributing the devcontainer CLI as a single executable application
allows to use it on systems where Node.js is not installed.
For example on immutable systems like Fedora Silverblue or
openSUSE MicroOS.