Changing the content of the config layer to an empty object
The content of the configuration layer of the feature manifest is specified as empty content (size 0).
This is because there is no content to reference in the config layer content of the feature manifest.

However, some registries that implement the OCI image manifest specification, such as AWS ECR, do not
support pushing empty content (size 0).
Therefore, this commit changes the content of the config layer of the feature manifestto indicate that
empty objects have no content to reference.