Changes the config layer of the feature manifest to a empty descriptor (#815)
* Change the config layer of the feature manifest to a empty descriptor

This commit changes the config layer of the feature manifest to a empty descriptor according to the OCI image manifest specification to to increase the available feature registry.

Some registries that implement the OCI image manifest specification, such as AWS ECR, do not support pushing empty content (size 0). However, the manifest may contain layers with no content. Therefore, the OCI Image Manifest specification specifies an empty descriptor of non-zero size to represent a layer with no content.

Currently, the devcontainer feature manifest specifies the config layer as an empty content (size 0) layer. If the devcontainer feature manifest conforms to this specification, it can also available registries that cannot publish empty content (size 0).

Specification: https://github.com/opencontainers/image-spec/blob/8f3820ccf8f65db8744e626df17fe8a64462aece/manifest.md#guidelines-for-artifact-usage

Note: If the config is set to an empty descriptor, the manifest media type must be defined in the artifactType.

* Revert "Change the config layer of the feature manifest to a empty descriptor"

This reverts commit 1b290f0c054cf5bd037c858f62ad58f2eda8cdf9.

* Changing the content of the config layer to an empty object

The content of the configuration layer of the feature manifest is specified as empty content (size 0).
This is because there is no content to reference in the config layer content of the feature manifest.

However, some registries that implement the OCI image manifest specification, such as AWS ECR, do not
support pushing empty content (size 0).
Therefore, this commit changes the content of the config layer of the feature manifestto indicate that
empty objects have no content to reference.

* Revert the test of push against the GitHub registry and fix/add the comments

* Change the feature that test depends on